requirements.txt
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
docutils==0.18.1
idna==3.3
importlib-metadata==4.11.3
jeepney==0.8.0
keyring==23.5.0
livia-core==0.1.dev78
mypy==0.931
mypy-extensions==0.4.3
numpy==1.21.5
opencv-python-headless==4.5.5.62
packaging==21.3
pkginfo==1.8.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyroma==3.2
PySide2==5.15.2
readme-renderer==35.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
SecretStorage==3.3.2
shiboken2==5.15.2
six==1.16.0
tomli==2.0.1
tqdm==4.64.0
twine==3.8.0
typed-ast==1.5.3
typing_extensions==4.2.0
urllib3==1.26.9
webencodings==0.5.1
zipp==3.8.0
